Fix remarks for `dpnp.empty()` #728 (#741)
* Rename usm_allocators_ext
* Add docstrings for functions in usm_allocators_ext.c
* Use ArrayModel as a model for dpnp.ndarray
* Update tests for models and typeof
