+++ numba_dppy/dpnp_ndarray/models.py
@@ -12,49 +12,14 @@
 # See the License for the specific language governing permissions and
 # limitations under the License.
 
-from numba import types
-from numba.core.datamodel import StructModel
-from numba.core.datamodel.models import ArrayModel
+from numba.core.datamodel.models import ArrayModel as dpnp_ndarray_Model
 from numba.extending import register_model
 
 from numba_dppy.target import spirv_data_model_manager
 
 from .types import dpnp_ndarray_Type
 
-# we reuse models.ArrayModel
-# it should contain all properties from __sycl_usm_array_interface__
-# __sycl_usm_array_interface__ and __array_interface__ are different
-# _Memory.__sycl_usm_array_interface__ has data, shape, strides, typestr, version, syclobj
-
-
-class dpnp_ndarray_Model(StructModel):
-    """Model for DPNP array.
-
-    It contains all items from models.ArrayModel and adding properties from
-    __sycl_usm_array_interface__: syclobj
-    """
-
-    def __init__(self, dmm, fe_type):
-        ndim = fe_type.ndim
-        members = [
-            # from models.ArrayModel
-            ("meminfo", types.MemInfoPointer(fe_type.dtype)),
-            ("parent", types.pyobject),
-            ("nitems", types.intp),
-            ("itemsize", types.intp),
-            ("data", types.CPointer(fe_type.dtype)),
-            # from __sycl_usm_array_interface__
-            # ("syclobj", types.pyobject),
-            # from models.ArrayModel
-            ("shape", types.UniTuple(types.intp, ndim)),
-            ("strides", types.UniTuple(types.intp, ndim)),
-        ]
-        super().__init__(dmm, fe_type, members)
-
-
 # This tells Numba to use the default Numpy ndarray data layout for
 # object of type dpnp.ndarray.
-# register_model(dpnp_ndarray_Type)(ArrayModel)
-
 register_model(dpnp_ndarray_Type)(dpnp_ndarray_Model)
-spirv_data_model_manager.register(dpnp_ndarray_Type, ArrayModel)
+spirv_data_model_manager.register(dpnp_ndarray_Type, dpnp_ndarray_Model)
