+++ numba_dppy/numpy_usm_shared.py
@@ -74,13 +74,13 @@ def dprint(*args):
 import dpctl
 from dpctl.memory import MemoryUSMShared
 
-import numba_dppy._usm_shared_allocator_ext
+import numba_dppy._usm_allocators_ext
 
 # Register the helper function in dppl_rt so that we can insert calls to them via llvmlite.
 for (
     py_name,
     c_address,
-) in numba_dppy._usm_shared_allocator_ext.c_helpers.items():
+) in numba_dppy._usm_allocators_ext.c_helpers.items():
     llb.add_symbol(py_name, c_address)
 
 
@@ -239,7 +239,7 @@ def is_usm_callback(obj):
         while isinstance(mobj, numba.core.runtime._nrt_python._MemInfo):
             ea = mobj.external_allocator
             dppl_rt_allocator = (
-                numba_dppy._usm_shared_allocator_ext.get_external_allocator()
+                numba_dppy._usm_allocators_ext.get_external_allocator()
             )
             dprint("Checking MemInfo:", ea)
             if ea == dppl_rt_allocator:
